+++ app/schemas/billing/payment.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from pydantic import BaseModel, ConfigDict, Field
 from typing import Optional
 from uuid import UUID
@@ -7,7 +8,7 @@ class PaymentBase(BaseModel):
     user_id: UUID
     subscription_id: UUID
     payment_method_id: UUID
-    amount: float
+    amount: Decimal
     currency: str
     status: str
 
@@ -15,7 +16,7 @@ class PaymentCreate(BaseModel):
     user_id: UUID
     subscription_id: UUID
     payment_method_id: UUID
-    amount: float
+    amount: Decimal
     currency: str = Field(..., max_length=255)
     status: str = Field(..., max_length=255)
 
@@ -23,7 +24,7 @@ class PaymentUpdate(BaseModel):
     user_id: Optional[UUID] = None
     subscription_id: Optional[UUID] = None
     payment_method_id: Optional[UUID] = None
-    amount: Optional[float] = None
+    amount: Optional[Decimal] = None
     currency: Optional[str] = None
     status: Optional[str] = None
     created_at: Optional[datetime] = None
